+++ vendor-prefixed/psr/http-client/src/ClientExceptionInterface.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * @license MIT
+ *
+ * Modified by AxePress Development using Strauss.
+ * @see https://github.com/BrianHenryIE/strauss
+ */
+
+namespace WPGraphQL\Login\Vendor\Psr\Http\Client;
+
+/**
+ * Every HTTP client related exception MUST implement this interface.
+ */
+interface ClientExceptionInterface extends \Throwable
+{
+}