+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to a modified version of [Semantic Versioning](./README
 
 ## Unreleased
 
+- dev: relocate Strauss dependencies to `vendor-prefixed`.
+- chore: update Strauss and Composer deps.
+
 ## [0.0.4] - 2022-02-09
 
 This release adds support for setting a WP Authentication Cookie on successful login, as well as compatibility with WPGraphQL for WooCommerce. It also fixes a handful of bugs, and backfills/refactors CI tests.